#11 refactoring
