+++ routes/user.js
@@ -4,14 +4,7 @@ const userCtrl = require("../controllers/user");
 const Validator = require("../middleware/validator");
 const ValidatorError = require("../middleware/validatorError");
 const auth = require("../middleware/auth");
-const passport = require("passport");
-require("../utils/kakaoPassport")(passport);
 require("dotenv").config();
-const kakaoKey = {
-    clientID: process.env.KAKAO_CLIENT_ID,
-    clientSecret: process.env.KAKAO_CLIENT_SECRET,
-    callbackURL: process.env.KAKAO_CALLBACK_URL
-};
 
 router.post("/join", Validator.join, ValidatorError.join, userCtrl.join);
 router.post("/login", Validator.login, ValidatorError.login, userCtrl.login);
@@ -28,22 +21,8 @@ router.post(
 );
 router.get("/logout", auth.verifyToken, userCtrl.logout);
 router.get("/auth", auth.verifyToken, userCtrl.auth);
-// test api
+
+// accessToken, refeshToken 재발급 과정 동작이 원활한지 테스트를 만들도록 함
 router.get("/dashboard", auth.verifyToken, userCtrl.dashboard);
-
-router.get("/kakao", passport.authenticate("kakao-login"));
-router.get(
-    "/auth/kakao/callback",
-    passport.authenticate("kakao-login", {
-        successRedirect: "/",
-        failureRedirect: "/api/auth/fail"
-    })
-);
-
-router.get("/test", (req, res) => {
-    const kakaoAuthUrl = `https://kauth.kakao.com/oauth/authorize?client_id=${kakaoKey.clientID}&redirect_uri=${kakaoKey.callbackURL}&response_type=code`;
-
-    return res.redirect(kakaoAuthUrl);
-});
 
 module.exports = router;
